antenna/JSON fortran compiles now. not working yet
Figured out how to get it to compile properly. Annoyingly I am having
to hardcode an -I and -L in the Makefile for now otherwise it can't
find the JSON parser. but whatever. segfaults in fitness_calc now
but I think that's because I've indexed something wrong in the twa loops