add code to connect adjacent branches - not tested yet
We want to consider the case where branches can exchange energy;
in principle I don't think it should make a difference to the results,
but some people have suggested that it does, and it's something a
reviewer might ask. next, write some code to generate pairs of antennae
that are identical except for allowing this transfer, and check the
resulting fitness values.