connected antenna works now - hopping rates in correct places
had to add a check when hopping across the "boundary" from the last
branch in the array to the first; there's four extra entries in there
which correspond to the RC and trap that you have to skip. manual
inspection of the K matrices via test_connected_kmat in tests.py shows
that the hopping rates are being places at the correct indices. nu_e
and efficiency values look identical, but test that next