rename phi_f -> phi_e. add class member assigment in tests
the fitness function in genetic algorithm takes an instance of Genome as
an argument and uses its nu_e member to calculate fitness, but this
value needs to be set by hand because the code that calculates it won't
necessarily know about the dataclass structure (i.e. if i hopefully
switch over to fortran soon). so for now just do it by hand in the test;
preferably write a wrapper function to do it though i guess?