fortran compiles, runs as shared lib, error in array lengths/char array?
i finally got it to be called from the python and actually work, maybe?
but i think there's an issue with the size of the character array, and
possibly the offset array as well. gonna make a small test antenna and
just work on that one first, in a separate test function