make cost into a variable for now, update various things to reflect
since we're gonna have to loop over cost parameter again multiple times,
i've made it into a parameter in the loop in main.py. a few things have
had to be changed to reflect this which can be changed back once we
settle on a value, e.g. fitness and selection functions. added in the
plot of averaged spectra vs. cost as well - there was a typo but i think
it should work now