copy best genome to separate variable instead of just best = pop[j]
i've been occasionally seeing weird antenna plots and i suddenly
realised a possible reason - best was simply assigned as pop[j], which
will just create a reference. now do a proper copy and only change it
when the maximum fitness increases. also sort out the julia to use the
JSON file and fix the latex text for the cyano pigments - can't put a
hyphen in a latex equation