add check on selection failing. various refactoring
add a check on the maximum fitness so that if the algorithm can't
generate a set of workable antennae it quits and starts the next run.
this involves adding some bools to avoid trying to plot averages etc.
for a run that never happened, but that is also done now. also reduce
the hist_snapshot to every 25 generations, remove some unused imports,
delete some old functions that were never used anywhere, fix a typo in
the JSON (r-PE and c-PE were both labelled r-PE in the text field; the
actual data wasn't affected)