add average antenna function to stats.py
new average_antenna function in stats.py reads in the averages and
writes out a set of parameters for an "averaged" antenna, from the final
population of each run, for us to draw later. also averages all of the
final averaged absorption spectra from each run, outputs the normalised
lineshape and plots the result. tested it with a toy run and it works
fine.