fix loops in main.py
i forgot that light.build returns a zip - once it's iterated over once,
it needs to be rebuilt, so you have to do that inside the outer loop.
sorted now. basically ready to run the final set of simulations with
efficiency weighting now, just need to add it in fitness function