first try at antenna-RC combined matrix. changes to pigment_data
not tested yet - the changes to pigment_data are to allow for two RCs at
the start of the pigment list. antenna_RC function in rc_toy.py is
written; need to add import statements and generate an antenna to couple
it to to test. also added an infinity to nnls.f90 - add this to
standalone repo for testing