overhauled GA completely. replacing stats and plots in progress
basically i've changed it so the GA is now defined in
genetic_algorithm.py as a dict of parameters and then just iterated over
for the crossover, mutation operators etc. added a simple set of
assertions too which can be called by `python genetic_algorithm.py`,
just checking that no illegal mutations or anything are done. works fine