fix GA on template instantiation
the GA was breaking for various arcane reasons when instantiated via
template - turns out this is because the dataclass does not convert
ints, floats, str or lists to numpy equivalents, which i thought it
would because i set them as the defaults. so now check type on template
instantiation and convert as necessary. runs and passes the assertion
checks now