fix template initialisation and copying of genomes
the numpy arrays weren't being initialised properly so for any genomes
that were template instantiated or copied, the arrays all pointed to the
same place. very annoying. also minor changes here and there related to
testing i'm doing, e.g. rho and aff removed from genome for now to test
effect of mutations and so on