remove RC pigments from fitness calc, change arguments for solvers
fitness calculation now only considers the size of the antenna, so that
we don't artificially disadvantage our genomes based on their RC type.
also change both RC_only and antenna_RC solvers to take the rc
type/genome and spectrum and then kwargs, just to make them a bit more
consistent. add the diffusion ratios for each RC type to this as well