aborted attempt to calc redox states on the fly
i tried a thing here to generate the set of states corresponding to the
trap, oxidised, reduced, etc. and then get the redox states from that,
without generating the tuples corresponding to each states. it doesn't
work, and i'm going to remove it and replace it with a switch so that we
can turn off calculating redox states entirely to save time. but i'm
leaving this commit as the half-finished attempt if i need it later