possible fix for k_b and new build_matrix function
there was an indexing issue in the antenna branch rates and also some
indentation errors in build_matrix which were causing the function to
return early before it had built the entire thing. stupid mistake. need
to check k_b over one more time and then run a long test of the old and
new solvers to check equality