refactoring! moved simulation to new file, argparse, etc
i've added simulation.py, which contains most of what was in main.py. it
takes a few arguments, the main one of which is a file containing the
spectrum. this is so i can submit array jobs on apocrita and run one
simulation per core. main.py (by default) now takes some arguments,
constructs the directory tree for the set of spectra we're looking it,
generates the spectra, saves them and then outputs a list of files
(this will go in the job script)